requirements.txt
indic-transliteration
Jinja2
python-docx
